Sql2017cu (#108)
* Trying to fix update

* Working better

* Fix 2017 CU23

- Improve update scripts. Just get checksum for 64

+++ sql-server-2017-cumulative-update/tools/chocolateyinstall.ps1
@@ -1,19 +1,21 @@
 ﻿$ErrorActionPreference = 'Stop';
 $toolsDir   = "$(Split-Path -parent $MyInvocation.MyCommand.Definition)"
 
-$url        = 'https://download.microsoft.com/download/C/4/F/C4F908C9-98ED-4E5F-88D5-7D6A5004AEBD/SQLServer2017-KB4577467-x64.exe'
-$checksum   = 'c20313b4f39afb8aaf14d6b100da8bb04295029116a7ce76b8f7c30b0499b154'
-$softwareName = 'Hotfix 3381 for SQL Server 2017*(KB4577467)*'
+$url        = 'https://download.microsoft.com/download/C/4/F/C4F908C9-98ED-4E5F-88D5-7D6A5004AEBD/SQLServer2017-KB5000685-x64.exe'
+$checksum   = 'dfa05d11624badf91b2d4ad03b59ce24ad1851ad94bba0a0c963d2275565007d'
+$softwareName = 'Hotfix 3381 for SQL Server 2017*(KB5000685)*'
 
+[bool] $runningAU = (Test-Path Function:\au_GetLatest)
 $filename = [IO.Path]::GetFileName($url)
 
 # Download like Install-ChocolateyPackage (so we can restart from cached download)
 $chocTempDir = $env:TEMP
+
 $tempDir = Join-Path $chocTempDir "$($env:chocolateyPackageName)"
 if ($env:chocolateyPackageVersion -ne $null) { $tempDir = Join-Path $tempDir "$($env:chocolateyPackageVersion)"; }
 $tempDir = $tempDir -replace '\\chocolatey\\chocolatey\\', '\chocolatey\'
 if (![System.IO.Directory]::Exists($tempDir)) { [System.IO.Directory]::CreateDirectory($tempDir) | Out-Null }
-$downloadFilePath = Join-Path $tempDir "$($packageName)Install.$fileType"
+$downloadFilePath = Join-Path $tempDir $filename
 
 $fullFilePath = Join-Path $toolsDir $filename
 
@@ -26,6 +28,10 @@ $packageArgs = @{
 }
 
 $filePath = Get-ChocolateyWebFile @packageArgs
+
+if (Test-Path Function:\au_GetLatest) {
+  return
+}
 
 # Copy into tools to keep for uninstall
 Copy-Item $filePath -Destination $fullFilePath
